% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{accuracy}
\alias{accuracy}
\title{Calculate accuracy based on comparing predicted and true labels}
\usage{
accuracy(pred, truth)
}
\arguments{
\item{pred}{A vector of predicted labels (0 or 1)}

\item{truth}{A vector of true labels (0 or 1)}
}
\value{
The accuracy of the model (a value between 0 and 1)
}
\description{
This function calculates the accuracy of a binary classification model based on the predicted and true labels.
}
\examples{
pred <- c(1, 0, 1, 0, 1)
truth <- c(1, 1, 0, 0, 1)
accuracy(pred, truth)
# Output: 0.6

}
