% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{adjR2}
\alias{adjR2}
\title{Calculate the adjusted r-squared statistic between predicted and ground truth continuous variables.}
\usage{
adjR2(pred, truth, n_p)
}
\arguments{
\item{pred}{The predicted values.}

\item{truth}{The ground truth values.}

\item{n_p}{The number of model parameters, excluding the intercept.}
}
\value{
The adjusted r-squared statistic.
}
\description{
Calculate the adjusted r-squared statistic between predicted and ground truth continuous variables.
}
\examples{
# Generate some example data
set.seed(123)
pred <- rnorm(100, 0, 1)
truth <- 0.5*pred + rnorm(100, 0, 0.5)

# Calculate the adjusted R-squared
adjR2(pred, truth, 1)

}
\seealso{
\code{\link{r2}}
}
