% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mat.R
\name{cor_mat}
\alias{cor_mat}
\title{A cor_mat function}
\usage{
cor_mat()
}
\arguments{
\item{x}{An individual year}
}
\description{
This function allows you to create a correlation matrix for all numeric variables in the data for the year.
}
\examples{
cor_mat(1950)
}
\keyword{Year,}
\keyword{correlation}
\keyword{matrix}
