% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{f1}
\alias{f1}
\title{Calculate the F1 score based on comparing predicted and true labels}
\usage{
f1(pred, truth)
}
\arguments{
\item{pred}{a binary vector of predicted labels}

\item{truth}{a binary vector of true labels}
}
\value{
the F1 score
}
\description{
Calculate the F1 score based on comparing predicted and true labels
}
\examples{
pred <- c(1, 0, 1, 1, 0)
truth <- c(0, 0, 1, 1, 1)
f1(pred, truth)
}
