% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{is_max}
\alias{is_max}
\title{Check if Values in a Vector are the Maximum Value}
\usage{
is_max(x, na.rm = T)
}
\arguments{
\item{x}{a vector to test}

\item{na.rm}{logical; should missing values (NA) be removed before calculating the maximum?}
}
\value{
a logical vector indicating whether the values in \code{x} are equal to the maximum value of \code{x}
}
\description{
This function checks if the values in a vector are equal to the maximum value
of the vector. The maximum value can be calculated with or without NA values.
}
\examples{
is_max(c(1, 2, 3, 1, 4, 5, 6, 7, 8, 8))
is_max(c(1, 2, 3, 1, 4, NA, 6, 7, 8, 8), na.rm = TRUE)
}
