% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_chisq}
\alias{log_likelihood_chisq}
\title{Calculate the log-likelihood of a chi-squared distribution}
\usage{
log_likelihood_chisq(x, df)
}
\arguments{
\item{x}{a numeric vector of observations}

\item{df}{the degrees of freedom of the chi-squared distribution}
}
\value{
a numeric value representing the log-likelihood of the observations given the
        degrees of freedom of the chi-squared distribution
}
\description{
Calculates the log-likelihood of an observation or vector of observations given the
degrees of freedom of a chi-squared distribution.
}
\examples{
log_likelihood_chisq(3.5, 2)
log_likelihood_chisq(c(2, 4, 6), 3)

}
