% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_f}
\alias{log_likelihood_f}
\title{Calculate the log-likelihood of an F distribution with given parameters}
\usage{
log_likelihood_f(x, df1, df2)
}
\arguments{
\item{x}{A numeric vector of observations}

\item{df1}{The degrees of freedom parameter for the numerator distribution}

\item{df2}{The degrees of freedom parameter for the denominator distribution}
}
\value{
The log-likelihood of the F distribution with the given parameters and
        the observed data
}
\description{
This function calculates the log-likelihood of an F distribution with given
degrees of freedom parameters df1 and df2, evaluated at a set of observations x.
}
\examples{
x <- rf(100, 5, 10)
log_likelihood_f(x, 5, 10)
}
