% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_norm}
\alias{log_likelihood_norm}
\title{Calculate the log-likelihood of a normal distribution}
\usage{
log_likelihood_norm(x, mean, sd)
}
\arguments{
\item{x}{a numeric vector of observations}

\item{mean}{the mean of the normal distribution}

\item{sd}{the standard deviation of the normal distribution}
}
\value{
a numeric value representing the log-likelihood of the observations given the
        parameters of the normal distribution
}
\description{
Calculates the log-likelihood of an observation or vector of observations given the
parameters of a normal distribution: the mean and standard deviation.
}
\examples{
log_likelihood_norm(3, 0, 1)
log_likelihood_norm(c(1, 2, 3), 0, 1)

}
