% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_t}
\alias{log_likelihood_t}
\title{Returns the log-likelihood of a sample x under the t-density}
\usage{
log_likelihood_t(x, df)
}
\arguments{
\item{x}{a numeric vector of sample data}

\item{df}{degrees of freedom}
}
\value{
the log-likelihood of the sample x under the t-density
}
\description{
Returns the log-likelihood of a sample x under the t-density
}
\details{
Computes the log-likelihood of the sample x under the t-density using the standard parameterization described in ?dt.
}
\examples{
# Generate a random sample from t-distribution
set.seed(123)
x <- rt(100, df = 5)

# Compute the log-likelihood of the sample under the t-density
log_likelihood_t(x, df = 5)

}
