% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_unif}
\alias{log_likelihood_unif}
\title{Calculate the log-likelihood of a uniform distribution}
\usage{
log_likelihood_unif(x, min, max)
}
\arguments{
\item{x}{a numeric vector of observations}

\item{min}{the minimum value of the uniform distribution}

\item{max}{the maximum value of the uniform distribution}
}
\value{
a numeric value representing the log-likelihood of the observations given the
        parameters of the uniform distribution
}
\description{
Calculates the log-likelihood of an observation or vector of observations given the
parameters of a uniform distribution: the minimum and maximum values.
}
\examples{
log_likelihood_unif(0.5, 0, 1)
log_likelihood_unif(c(0.1, 0.2, 0.3), 0, 1)

}
