% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{minimum_n_per_group}
\alias{minimum_n_per_group}
\title{Calculate Minimum Sample Size per Group for t-Test at Given Power}
\usage{
minimum_n_per_group(d, power = 0.8)
}
\arguments{
\item{d}{expected Cohen’s d.}

\item{power}{desired power}
}
\value{
minimum sample size per group required to detect the given effect size with the given power level and significance level
}
\description{
This function calculates the minimum sample size per group required to detect a given effect size with a t-test
at a given power level and significance level.
}
\examples{
minimum_n_per_group(0.5, power = 0.9)
}
