% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{precision}
\alias{precision}
\title{Calculate Precision}
\usage{
precision(pred, truth)
}
\arguments{
\item{pred}{A vector of predicted labels.}

\item{truth}{A vector of truth labels.}
}
\value{
The precision value.
}
\description{
Calculates the precision based on comparing predicted and truth labels.
}
\examples{
precision(c(1,1,0,1),c(1,0,0,1))
}
