% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{r2}
\alias{r2}
\title{Calculate the R-squared statistic between predicted and ground truth continuous variables}
\usage{
r2(pred, truth)
}
\arguments{
\item{pred}{predicted values}

\item{truth}{ground truth values}
}
\value{
the R-squared statistic
}
\description{
Calculate the R-squared statistic between predicted and ground truth continuous variables
}
\examples{
r2(c(1, 2, 3), c(1, 3, 5))

}
