% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{recall}
\alias{recall}
\title{Calculate recall (same as sensitivity) based on comparing predicted and training labels}
\usage{
recall(pred, truth)
}
\arguments{
\item{pred}{A vector of predicted labels (0/1)}

\item{truth}{A vector of true labels (0/1)}
}
\value{
The recall (sensitivity) of the prediction
}
\description{
Calculate recall (same as sensitivity) based on comparing predicted and training labels
}
\examples{
truth <- c(1, 1, 0, 0, 1, 0, 1, 1, 1, 0)
pred <- c(1, 0, 0, 1, 1, 0, 1, 1, 1, 0)
recall(pred, truth)

}
