% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{rep_mat}
\alias{rep_mat}
\title{Replicate matrix rows or columns}
\usage{
rep_mat(x, M = 1, N = 1)
}
\arguments{
\item{x}{a matrix or data.frame object}

\item{M}{an integer specifying the number of times to replicate the rows (default is 1)}

\item{N}{an integer specifying the number of times to replicate the columns (default is 1)}
}
\value{
a matrix with replicated rows and/or columns
}
\description{
Creates a matrix by replicating the rows or columns of a given matrix or data.frame
M (N) times.
}
\examples{
x <- matrix(1:4, ncol = 2)
rep_mat(x, 2, 1)
rep_mat(x, 1, 2)

}
