% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{sensitivity}
\alias{sensitivity}
\title{Calculate the sensitivity of a binary classifier.}
\usage{
sensitivity(pred, truth)
}
\arguments{
\item{pred}{A binary vector of predicted labels.}

\item{truth}{A binary vector of actual labels.}
}
\value{
The sensitivity of the binary classifier.
}
\description{
Sensitivity measures the proportion of true positive cases that were correctly identified
as positive by the binary classifier. It is calculated as the ratio of true positives to
all actual positives.
}
\examples{
pred <- c(1,0,1,1,0,0,1)
truth <- c(1,0,1,0,0,0,0)
sensitivity(pred, truth)
}
