% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{specificity}
\alias{specificity}
\title{Calculate specificity based on predicted and true labels}
\usage{
specificity(pred, truth)
}
\arguments{
\item{pred}{A vector of predicted binary values.}

\item{truth}{A vector of true binary values.}
}
\value{
A numeric value representing the specificity.
}
\description{
This function takes two vectors of binary values, the predicted and true labels, and returns the specificity, i.e., the proportion of negative cases correctly identified as negative.
}
\examples{
specificity(c(1,1,0,0), c(1,0,1,0))

}
