% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_good.R
\name{who_good}
\alias{who_good}
\title{A who_good function}
\usage{
who_good()
}
\arguments{
\item{x}{An individual year}
}
\description{
This function allows you to find out the best player of the year based on his total sum of score.
}
\examples{
who_good(1950)
}
\keyword{PTS}
\keyword{Player,}
\keyword{Year,}
