% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_old.R
\name{who_old}
\alias{who_old}
\title{A who_old function}
\usage{
who_old()
}
\arguments{
\item{x}{An individual year}
}
\description{
This function allows you to find out the oldest player of the year.
}
\examples{
who_old(1950)
}
\keyword{Age}
\keyword{Player,}
\keyword{Year,}
