% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{is_max}
\alias{is_max}
\title{Test if elements are maximum of a vector}
\usage{
is_max(x, na.rm = T)
}
\arguments{
\item{x}{an atomic vector of numeric, integer, or logical values}

\item{na.rm}{a logical value indicating whether missing values should be removed (default is TRUE)}
}
\value{
a logical vector of the same length as x
}
\description{
Returns a logical vector indicating whether the elements of the input vector are
equal to its maximum value.
}
\examples{
is_max(c(1,2,3,1))
is_max(c(1,2,3,1), na.rm = FALSE)

}
