% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_t}
\alias{log_likelihood_t}
\title{Calculate Log-Likelihood for t-Distribution}
\usage{
log_likelihood_t(x, df)
}
\arguments{
\item{x}{a numeric vector of data}

\item{df}{degrees of freedom of the t-distribution}
}
\value{
the log-likelihood of the data assuming a t-distribution with the given degrees of freedom
}
\description{
This function calculates the log-likelihood of a t-distribution with a given degrees of freedom parameter
and a given set of data.
}
\examples{
log_likelihood_t(c(1, 2, 3, 4, 5), df = 3)
}
