% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{minimum_n_per_group}
\alias{minimum_n_per_group}
\title{Calculate minimum n per group for a two-sample t-test}
\usage{
minimum_n_per_group(d, power = 0.8)
}
\arguments{
\item{d}{Expected Cohen's d}

\item{power}{Desired statistical power}
}
\value{
Minimum n per group needed for a two-sample t-tesinslibrl
}
\description{
Calculate minimum n per group for a two-sample t-test
}
\examples{
minimum_n_per_group(0.5)
}
